Python tests for TirePressureMonitoringSystem
I removed count in Alarm since I didn't see any use for it. The Sensor
is injected to allow for mocks. It would also make it possible for the
Alarm to read any sensor that implements a pressure_psi_value method.
The name of the method in the Sensor was also shorted to just read
pressure_psi_value since the previous name revealed too much of the
implementation. The Alarm just needs to know the psi value, it doesn't
need to know that it has to pop it from somewhere.