Added initial tests for turn_ticket.py
Something bothers me about this code. For now, the tests are just for
the original code, but my intention is to refactor the code for the
following reasons:
There is a tight coupling between the dispenser and the sequence
generator. The sequence generator could be made pluggable by injecting a
duck into the dispenser. There is a problem with this in dynamic
languages. If the real production code's interface changes, then the
tests won't reflect it if we use doubles or mocks to test it. So I want
to find a way to make sure this doesn't happen.
Second, TurnNumber seems a little excesive since we are only returning a
number, I don't really see the need for returning an object.